+++ client/src/pages/Home.css
@@ -22,11 +22,15 @@
   color: #aaaaaa89;
 }
 
-.banners div span {
+.banners div a {
+  cursor: pointer;
   font-size: 2rem;
   padding: 0.2rem;
+  padding-left: 0.9rem;
 }
 
-.banners div span:hover {
+.banners div a:hover,
+.activeCl {
   color: black;
+  font-size: 2.5rem;
 }
