+++ client/src/pages/Home.js
@@ -5,17 +5,10 @@ import "./Home.css";
 import Category from "../components/Category";
 import Carousel from "../components/Carousel";
 
+let counter = 1;
+
 function Home() {
-  let [bannerData, setBannerData] = useState([]);
   let [categoriesData, setCategoriesData] = useState([]);
-  let [bannerImage, setBannerImage] = useState([]);
-  useEffect(() => {
-    fetch("http://localhost:5000/banners")
-      .then((data) => data.json())
-      .then((bannerdata) => {
-        setBannerData(bannerdata);
-      });
-  }, []);
 
   useEffect(() => {
     fetch("http://localhost:5000/categories")
@@ -24,25 +17,6 @@ function Home() {
         setCategoriesData(categoriesdata);
       });
   }, []);
-
-  let intervel;
-  let counter = 0;
-
-  let setCounter = (ind) => {
-    console.log(ind);
-    counter = ind;
-  };
-
-  useEffect(() => {
-    intervel = setInterval(() => {
-      setBannerImage([
-        bannerData[counter % 5]?.bannerImageUrl,
-        bannerData[counter++ % 5]?.bannerImageAlt,
-      ]);
-    }, 2000);
-
-    return () => clearInterval(intervel);
-  }, [bannerData]);
 
   let AllCategories = categoriesData.map((cat) => {
     if (cat.enabled)
@@ -59,21 +33,9 @@ function Home() {
       );
   });
 
-  let imageMover = bannerData.map((ban, index) => {
-    // if (cat.enabled)
-    return (
-      <span index={index} key={ban.id}>
-        &#8226;
-      </span>
-    );
-  });
-
   return (
     <article className="homepage-container">
-      <section className="banners">
-        <img src={bannerImage[0]} alt={bannerImage[1]} />
-        <div>{imageMover}</div>
-      </section>
+      <Carousel></Carousel>
 
       {AllCategories}
     </article>
